requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
opencv-python==4.8.1.78
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
Pillow==10.1.0
SQLAlchemy==2.0.23
